+++ crabml-core/src/tensor/mod.rs
@@ -2,5 +2,5 @@ pub mod cpu;
 pub mod strider;
 pub mod tensor;
 
-pub use cpu::arithmetic;
+pub use cpu::compute;
 pub use cpu::CpuTensor;
