+++ crabml-llama2/src/llama2.rs
@@ -5,15 +5,15 @@ use crabml::error::ErrorKind;
 use crabml::error::Result;
 use crabml::gguf::GGUFFile;
 use crabml::gguf::GGUFMetadata;
-use crabml::tensor::arithmetic::add_inplace;
-use crabml::tensor::arithmetic::batch_matmul;
-use crabml::tensor::arithmetic::div_scalar_inplace;
-use crabml::tensor::arithmetic::matmul;
-use crabml::tensor::arithmetic::mul_inplace;
-use crabml::tensor::arithmetic::rms_norm_inplace;
-use crabml::tensor::arithmetic::rope_inplace;
-use crabml::tensor::arithmetic::silu_inplace;
-use crabml::tensor::arithmetic::softmax_inplace;
+use crabml::tensor::compute::add_inplace;
+use crabml::tensor::compute::batch_matmul;
+use crabml::tensor::compute::div_scalar_inplace;
+use crabml::tensor::compute::matmul;
+use crabml::tensor::compute::mul_inplace;
+use crabml::tensor::compute::rms_norm_inplace;
+use crabml::tensor::compute::rope_inplace;
+use crabml::tensor::compute::silu_inplace;
+use crabml::tensor::compute::softmax_inplace;
 use crabml::tensor::CpuTensor;
 use std::ops::AddAssign;
 use std::time::Duration;
